add documentation to all exported methods. refactor cmd_runner to testhelpers
