+++ api/http/access_token.go
@@ -15,6 +15,7 @@ type accessTokenResponse struct {
 	Token string `json:"access_token"`
 }
 
+// GetAccessToken call to sign-in endpoint and get an access_token to use in later API calls.
 func GetAccessToken(cfg *config.Config, server *config.ServerConfig) (string, error) {
 	client := &http.Client{}
 	url := fmt.Sprintf("%s/api/v1/auth/token/signin", server.URL)
@@ -26,14 +27,14 @@ func GetAccessToken(cfg *config.Config, server *config.ServerConfig) (string, er
 
 	req, err := http.NewRequestWithContext(ctx, "POST", url, postData)
 	if err != nil {
-		return "", fmt.Errorf("error creating request call: %s", err)
+		return "", fmt.Errorf("error creating request call: %s", err) //nolint:goerr113
 	}
 
 	req.Header.Set("Content-Type", "application/json")
 
 	r, err := client.Do(req)
 	if err != nil {
-		return "", fmt.Errorf("error calling access token URL: %s", err)
+		return "", fmt.Errorf("error calling access token URL: %s", err) //nolint:goerr113
 	}
 	defer r.Body.Close()
 
@@ -45,7 +46,7 @@ func GetAccessToken(cfg *config.Config, server *config.ServerConfig) (string, er
 
 	err = json.NewDecoder(r.Body).Decode(&t)
 	if err != nil {
-		return "", fmt.Errorf("error decoding access token response: %s", err)
+		return "", fmt.Errorf("error decoding access token response: %s", err) //nolint:goerr113
 	}
 
 	return t.Token, nil
