+++ cmdutil/factory.go
@@ -9,6 +9,7 @@ import (
 	"github.com/konstellation-io/kli/internal/logger"
 )
 
+// CmdFactory interface to define all methods needed during commands creation.
 type CmdFactory interface {
 	IOStreams() *IOStreams
 	Config() *config.Config
@@ -16,6 +17,7 @@ type CmdFactory interface {
 	ServerClient(string) (api.ServerClienter, error)
 }
 
+// Factory contains all data needed during commands execution.
 type Factory struct {
 	appVersion string
 	ioStreams  *IOStreams
@@ -23,6 +25,7 @@ type Factory struct {
 	logger     logger.Logger
 }
 
+// NewCmdFactory returns a new Factory object used during commands execution.
 func NewCmdFactory(appVersion string) *Factory {
 	io := System()
 
@@ -34,18 +37,22 @@ func NewCmdFactory(appVersion string) *Factory {
 	}
 }
 
+// IOStreams access to IOStreams object.
 func (f *Factory) IOStreams() *IOStreams {
 	return f.ioStreams
 }
 
+// Config access to Config object.
 func (f *Factory) Config() *config.Config {
 	return f.cfg
 }
 
+// Logger access to Logger object.
 func (f *Factory) Logger() logger.Logger {
 	return f.logger
 }
 
+// ServerClient generates a new ServerClient specific for the given server name.
 func (f *Factory) ServerClient(serverName string) (api.ServerClienter, error) {
 	server := f.cfg.GetByServerName(serverName)
 	if server == nil {
