+++ internal/errors/errors.go
@@ -5,6 +5,8 @@ import (
 )
 
 var (
-	ErrUnknownServerName   = errors.New("unknown server name")
+	// ErrUnknownServerName used when server name is unknown.
+	ErrUnknownServerName = errors.New("unknown server name")
+	// ErrServerAlreadyExists used when adding a server with duplicated name.
 	ErrServerAlreadyExists = errors.New("server already exists")
 )
