+++ pkg/cmd/kre/runtime/runtime.go
@@ -8,6 +8,7 @@ import (
 	"github.com/konstellation-io/kli/pkg/cmd/kre/runtime/list"
 )
 
+// NewRuntimeCmd creates a new command to handle 'runtime' subcommands.
 func NewRuntimeCmd(f cmdutil.CmdFactory) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "runtime",
