+++ text/clean.go
@@ -10,14 +10,17 @@ var (
 	wt = regexp.MustCompile(`\t+`)
 )
 
+// Normalize return clean and lower case text.
 func Normalize(s string) string {
 	return strings.ToLower(Sanitize(s))
 }
 
+// Sanitize remove multiple spaces and return trimmed string.
 func Sanitize(s string) string {
 	return ws.ReplaceAllString(strings.TrimSpace(s), " ")
 }
 
+// LinesTrim remove multiples spaces and trim each line.
 func LinesTrim(s string) string {
 	lines := strings.Split(s, "\n")
 	for i, l := range lines {
